Start to clean the example repo - and create new ini file in python
