+++ examples/perfect/so_parameters.py
