--- examples/simple_parameters.ini
@@ -1,150 +1,0 @@
-[s4cmb]
-## Parameter file for a fake experiment.
-## The format is <name_in_small_caps> = <value> <letter>
-## where is used to determine the type of the value:
-## S(tring), F(loat), I(nteger), B(ool), N(one).
-
-## Run ID
-name_instrument = fake S
-
-#####################################################################
-######################### Input sky #################################
-#####################################################################
-## can be two things:
-##  1) Input fits file containing the sky maps to scan (maps will be loaded).
-##  2) CAMB lensed cl file (.dat) containing lensed power spectra with
-##      order ell, TT, EE, BB, TE (maps will be created on-the-fly).
-## /!\ Option 2) is more for test purposes /!\
-input_filename = s4cmb/data/test_data_set_lensedCls.dat S
-
-## Set do_pol to False if you want to load only intensity map.
-do_pol = True B
-
-## If input_filename is a CAMB lensed cl file, the code generate maps at
-## a resolution nside_in and convolved with a beam having this FWHM_in [arcmin].
-## No effect if you provide fits file with maps instead of the CAMB file.
-## /!\ this is more for test purposes /!\
-fwhm_in = 0.0 F
-nside_in = 512 I
-map_seed = 547389 I
-
-# Remove either I, Q or U to remove possible leakages
-no_ileak = False B
-no_quleak = False B
-
-# Set it to True if you are reading a map in Galactic coordinate.
-# (Planck maps for example).
-ext_map_gal = False B
-
-#####################################################################
-######################### Instrument ################################
-#####################################################################
-## creates the data used to model the instrument (see s4cmb/instrument.py).
-## Focal plane parameters: Number of crate plate, number of MUX board per crate,
-## number of SQUID per MUX and number of pair of bolometers per SQUID.
-ncrate = 1 I
-ndfmux_per_crate = 1 I
-nsquid_per_mux = 1 I
-npair_per_squid = 4 I
-
-## The size of the focal plane [cm].
-fp_size = 60. F
-
-## Beam parameters of the bolometers.
-## Full Width Half Maximum [arcmin].
-fwhm = 3.5 F
-
-## Seed used to generate angle of rotation of beam axes.
-beam_seed = 58347 I
-
-## Diameter of the focal plane [degree].
-projected_fp_size = 3. F
-
-# Pointing model parameters of the telescope.
-pm_name = 5params S
-
-# Polarisation angle of the bolometers and HWP. Choose among CRHWP or stepped.
-type_hwp = CRHWP S
-
-## HWP frequency [Hz] (only if CRHWP is used).
-freq_hwp = 2. F
-
-## starting position of the HWP if type_hwp=CRHWP is used
-## or the daily step change of the HWP if type_hwp=stepped is used.
-## [deg].
-angle_hwp = 0. F
-
-## Time-domain Noise level for the whole array.
-## Default are noiseless simulations (i.e. = None).
-## If you want to inject white noise on-the-fly, then specify a
-## noise level in [u]K.sqrt(s). Careful the units has to be the same as
-## the input map! Note also that it corresponds to the polarisation level.
-array_noise_level = None N
-
-## Seed used to generate random numbers to simulate noise.
-## From this single seed, we generate a list of seeds
-## for all detectors. Has an effect only if array_noise_level is provided.
-array_noise_seed = 487587 I
-
-#####################################################################
-######################### Scanning strategy #########################
-#####################################################################
-# A scanning strategy consists in defining the site of observation
-# on earth for which we will make the observation, the region
-# of the sky to observe, and the schedule of observations.
-# (see s4cmb/scanning_strategy.py)
-
-## Name of a pre-defined scanning strategy to define the boundaries
-## of the scan: elevation, azimuth, and time. Only available for the:
-## moment name_strategy = deep_patch.
-name_strategy = deep_patch S
-
-## Number of CES to run. Should not exceed the number of CES contained in
-## the defined scanning strategy set in name_strategy.
-nCES = 12 I
-
-## Starting date for observations (The format is: YYYY/M/D HH:MM:SS)
-start_date = 2013/1/1 00:00:00 S
-
-## Position of the telescope on Earth
-telescope_longitude = -67:46.816 S
-telescope_latitude = -22:56.396 S
-
-## Elevation (Alt) of the telescope [meter]
-telescope_elevation = 5200. F
-
-## Sampling frequency at which time samples are recorded [Hz]
-sampling_freq = 8. F
-
-## Azimuth speed of the telescope [deg/s]
-sky_speed = 0.4 F
-
-## File containing time correction to UTC
-ut1utc_fn = s4cmb/data/ut1utc.ephem S
-
-#####################################################################
-######################### Output sky maps ###########################
-#####################################################################
-## Choose the projection of the output maps: healpix of flat
-projection = healpix S
-
-## If projection=healpix, choose the nside
-nside_out = 512 I
-
-## If projection=flat, choose the pixel_size
-pixel_size = None N
-
-## Width of the output map in degree.
-## If width < len(input scan), you will get a cropped map.
-## if width > len(input scan), you will have your entire scan + plenty of zeros.
-## A choice has to be made.
-width = 20. F
-
-#####################################################################
-######################### Misc ######################################
-#####################################################################
-## language use to perform core computations.
-language = fortran S
-
-## Verbose mode.
-verbose = False B
